# Perf find in parallel