--- a/.gitignore
+++ b/.gitignore
@@ -1,1 +1,3 @@
 bin
+tmp
+.make