Running `helpmakego` in large makefiles (20 targets - each one a large go code base) gets pretty slow. I am seeing 2-4s to parse the Makefile (generating all targets).

This change moves the package parsing code from a stack-based depth first search to a fully parallel traversal.


This change leads to a 2-3x improvement in runtime performance on our current benchmark:

```
make benchmark                 
hyperfine --warmup 5 --command-name main 'tmp/helpmakego-main/bin/helpmakego tmp/pulumi/pkg/cmd/pulumi' --command-name current 'bin/helpmakego tmp/pulumi/pkg/cmd/pulumi'
Benchmark #1: main
  Time (mean ± σ):      73.5 ms ±   1.6 ms    [User: 104.2 ms, System: 27.0 ms]
  Range (min … max):    71.2 ms …  79.3 ms    34 runs
 
Benchmark #2: current
  Time (mean ± σ):      26.2 ms ±   1.4 ms    [User: 117.1 ms, System: 60.1 ms]
  Range (min … max):    22.7 ms …  30.3 ms    73 runs
 
Summary
  'current' ran
    2.80 ± 0.16 times faster than 'main'
```

I have observed the same change in multiple other repos.